move es6 rules into separate file
